requirements.txt
python-telegram-bot>=20.0a0
gspread>=3.7.0
email-validator>=1.1.2